Add SourceFingerprint based on path, check_name, and source
Currently, the default fingerprint for an issue is computed by creating an MD5
digest based on the path of the file which introduced the issue and the check
name for the issue emitted.

The result of this fingerprint for many engines is that issues emitted with the
same check name for the same file are given the same fingerprint. On
codeclimate.com, we only show issues with unique fingerprints, so even if 100
issues were emitted for a file, if all issues share the same check name and
thus the same fingerprint, we only show one issue.

Additionally, a fixed or introduced issue is often ignored if there are current
issues with the same fingerprint already on the default branch.

This PR introduces a new fingerprint to address these issues, first added to a
[couple][markdownlint] [engines][shellcheck], that incorporates relevant source
code into the fingerprint. This implementation allows us to enable the new
fingerprinting behavior with an environment variable so we can roll it out
gradually.

[markdownlint]: https://github.com/jpignata/codeclimate-markdownlint
[shellcheck]: https://github.com/filib/codeclimate-shellcheck